refactor: 누락된 components 폴더 추가
